Large overhall: Reintroduced Jastrow factors. Only one implemented is IdentityJastrow which equates to 1. Renamed Amplitude to SlaterDeterminant for clarity.

+++ Lattice.h
@@ -11,10 +11,10 @@ class Edge;
 class Vertex {
     public:
         size_t idx;
-        std::vector<int> uc; //!< Unit cell coordinates
+        std::vector<size_t> uc; //!< Unit cell coordinates
         std::vector<double> pos; //!< Position within unit cell
         std::vector<Edge*> edges;
-        Vertex(size_t idx, const std::vector<int>& uc, const std::vector<double>& pos)
+        Vertex(size_t idx, const std::vector<size_t>& uc, const std::vector<double>& pos)
             : idx(idx), uc(uc), pos(pos) 
         {}
         ~Vertex()
